improved Windows build notes and CMakeLists. added Windows installer project. fixed build error.

+++ src/gui/SendMessageFrame.cpp
@@ -159,7 +159,7 @@ void SendMessageFrame::sendClicked() {
       return;
     }
 
-    transfers.append({address.toStdString(), MESSAGE_AMOUNT});
+    transfers.append({address.toStdString(), (int64_t)MESSAGE_AMOUNT});
     messages.append({messageString.toStdString(), address.toStdString()});
   }
